pyproject.toml
[project]
name = "arxivsearcher"
version = "0.1.0"
description = ""
authors = [
    {name = "Celia ZHANG",email = "celiazhang349@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain (>=0.3.22,<0.4.0)",
    "arxiv",
    "langchain",
    "langchain-community",
    "langchain-huggingface",
    "langchain_chroma",
    "chromadb",
    "tiktoken",
    "sentence-transformers",
    "streamlit",
    "google-cloud-storage"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
